Performance boost. Object caching in different plugins

+++ packages/Db/Memcache/Managers/MemcacheWrapper.class.php
@@ -8,19 +8,35 @@ class MemcacheWrapper {
 	const MEMCACHE_DEFAULT = null;
 
 	/**
+	 *
+	 * @var type Config
+	 */
+	protected $config = null;
+
+	/**
 	 * @var Memcache
 	 */
-	private $memcache = null;
+	protected $memcache = null;
+
+	/**
+	 *
+	 * @var boolean 
+	 */
+	protected $isEnabled = false;
 
 	/**
 	 * Class constructor
 	 *
-	 * @param $memcache_host, $memcache_port
+	 * @param Config $config
 	 */
-	public function __construct($memcache_host = '127.0.0.1', $memcache_port = 11211) {
-		if (is_null($this->memcache) || (!is_object($this->memcache))) {
+	public function __construct(Config $config) {
+		$this->config = $config;
+
+		$this->isEnabled = $this->config->enabled;
+
+		if ($this->isEnabled) {
 			$this->memcache = new Memcache();
-			if (!$this->memcache->pconnect($memcache_host, $memcache_port)) {
+			if (!$this->memcache->pconnect($this->config->host, $this->config->port)) {
 				throw new MemcacheException("Error in object initialization", 2);
 			}
 		}
@@ -43,11 +59,14 @@ class MemcacheWrapper {
 	 * @param mixed $value
 	 * @param int $expire the number of seconds after which cached item expires
 	 * @param int $flag
-	 * @return bool ture if successful false otherwise
+	 * @return bool true if successful false otherwise
 	 */
 	public function set($key, $value, $expire = 0, $flag = 0) {
 		if (empty($key)) {
 			throw new InvalidArgumentException("\$key can't be empty");
+		}
+		if (!$this->isEnabled) {
+			return false;
 		}
 
 		if (!$this->memcache->set($key, $value, $flag, $expire)) {
@@ -65,7 +84,79 @@ class MemcacheWrapper {
 		if (empty($key)) {
 			throw new InvalidArgumentException("\$key is empty");
 		}
+
+		if (!$this->isEnabled) {
+			return false;
+		}
+
 		return $this->memcache->get($key);
+	}
+
+	/**
+	 * Helper function for caching objects
+	 * 
+	 * @param string $tag
+	 * @param mixed $value
+	 * @param int $cacheMinutes
+	 */
+	public function setObject($tag, $name, $value, $cacheMinutes = self::MEMCACHE_UNLIMITED) {
+		if ($cacheMinutes === self::MEMCACHE_OFF) {
+			return;
+		}
+		if (!$this->isEnabled) {
+			return false;
+		}
+		if(!empty($name)){
+			$name = ':' . $name;
+		}
+		
+		$key = $this->getNamespacedKey($tag) . $name;
+
+		$this->set($key, $value, $this->getExpirationByCacheMinutes($cacheMinutes));
+	}
+
+	/**
+	 * Helper function for getting cached objects
+	 * @param string $tag
+	 * @return mixed
+	 */
+	public function getObject($tag, $name = "") {
+		if (!$this->isEnabled) {
+			return false;
+		}
+		if(!empty($name)){
+			$name = ':' . $name;
+		}
+		
+		$key = $this->getNamespacedKey($tag) . $name;
+
+		return $this->get($key);
+	}
+
+	/**
+	 * Delete object from cache
+	 * 
+	 * @param string $tag
+	 * @return boolean
+	 */
+	public function deleteObject($tag) {
+		if (!$this->isEnabled) {
+			return false;
+		}
+		$key = $this->getNamespacedKey($tag);
+
+		return $this->delete($key);
+	}
+
+	public function getExpirationByCacheMinutes($cacheMinutes) {
+		if ($cacheMinutes > 0) {
+			$expire = time() + ($cacheMinutes * 60);
+		}
+		elseif ($cacheMinutes == self::MEMCACHE_UNLIMITED) {
+			$expire = 0;
+		}
+
+		return $expire;
 	}
 
 	/**
@@ -78,6 +169,10 @@ class MemcacheWrapper {
 		if (empty($key)) {
 			throw new InvalidArgumentException("\$key is empty");
 		}
+		if (!$this->isEnabled) {
+			return false;
+		}
+
 		return $this->memcache->increment($key);
 	}
 
@@ -91,6 +186,10 @@ class MemcacheWrapper {
 		if (empty($key)) {
 			throw new InvalidArgumentException("\$key is empty");
 		}
+		if (!$this->isEnabled) {
+			return false;
+		}
+
 		return $this->memcache->decrement($key);
 	}
 
@@ -98,6 +197,9 @@ class MemcacheWrapper {
 	 * Returns various server statistics
 	 */
 	public function getServerStats() {
+		if (!$this->isEnabled) {
+			return false;
+		}
 		return $this->memcache->getExtendedStats();
 	}
 
@@ -105,6 +207,9 @@ class MemcacheWrapper {
 	 * Clear all save items in Memcache
 	 */
 	public function clearAllItems() {
+		if (!$this->isEnabled) {
+			return false;
+		}
 		$this->memcache->flush();
 	}
 
@@ -116,6 +221,10 @@ class MemcacheWrapper {
 		if (empty($key)) {
 			throw new InvalidArgumentException("\$key is empty");
 		}
+		if (!$this->isEnabled) {
+			return false;
+		}
+
 		return $this->memcache->delete($key);
 	}
 
@@ -124,18 +233,22 @@ class MemcacheWrapper {
 	 * @return array
 	 */
 	public function getKeysList() {
+		if (!$this->isEnabled) {
+			return false;
+		}
+
 		$list = array();
 		$allSlabs = $this->memcache->getExtendedStats('slabs');
 		$items = $this->memcache->getExtendedStats('items');
 		foreach ($allSlabs as $server => $slabs) {
-			foreach ($slabs AS $slabId => $slabMeta) {
+			foreach ($slabs as $slabId => $slabMeta) {
 				if (!is_numeric($slabId)) {
 					continue;
 				}
 				$cdump = $this->memcache->getExtendedStats('cachedump', (int) $slabId, 1000000);
-				foreach ($cdump AS $server => $entries) {
+				foreach ($cdump as $server => $entries) {
 					if ($entries) {
-						foreach ($entries AS $eName => $eData) {
+						foreach ($entries as $eName => $eData) {
 							array_push($list, $eName);
 						}
 					}
@@ -148,8 +261,6 @@ class MemcacheWrapper {
 	}
 
 	public function getNamespacedKey($tag = null) {
-		$key = ConfigManager::getConfig("Db", "Memcache")->AuxConfig->keyPrefix . ":";
-
 		$key = $this->getKeyBeginning($tag);
 
 		$version = $this->get($this->getKeyBeginning($tag, true));
@@ -164,6 +275,9 @@ class MemcacheWrapper {
 	public function invalidateCacheByTag($tag = null) {
 		if (empty($tag)) {
 			throw new InvalidArgumentException("\$tag should be non empty string");
+		}
+		if (!$this->isEnabled) {
+			return false;
 		}
 
 		$versionKey = $this->getKeyBeginning($tag, true);
@@ -182,37 +296,37 @@ class MemcacheWrapper {
 		return $idTag;
 	}
 
-	public function deleteKeys($classes, $customGlobalPrefix = null){
-	       if(!is_array($classes)){
-		       $classes = array($classes);
-	       }
+	public function deleteKeys($classes, $globalPrefix = null) {
+		if (!$this->isEnabled) {
+			return false;
+		}
+		if (!is_array($classes)) {
+			$classes = array($classes);
+		}
 
-	       if($customGlobalPrefix !== null){
-		       $globalPrefix = $customGlobalPrefix;
-	       }
-	       else{
-		       $globalPrefix = ConfigManager::getConfig("Db", "Memcache")->AuxConfig->keyPrefix;
-	       }
+		if ($globalPrefix === null) {
+			$globalPrefix = $this->config->keyPrefix;
+		}
 
-	       $list = $this->getKeysList();
+		$list = $this->getKeysList();
 
-	       $count = 0;
-	       foreach ($list as $key){
-		       $array = explode(":", $key);
-		       if(count($array) >= 2){
-			       if($array[0] == $globalPrefix and in_array($array[1], $classes)){
-				       if($this->delete($key)){
-					       $count++;
-				       }
-			       }
-		       }
-	       }
+		$count = 0;
+		foreach ($list as $key) {
+			$array = explode(":", $key);
+			if (count($array) >= 2) {
+				if ($array[0] == $globalPrefix and in_array($array[1], $classes)) {
+					if ($this->delete($key)) {
+						$count++;
+					}
+				}
+			}
+		}
 
-	       return $count;
+		return $count;
 	}
 
 	protected function getKeyBeginning($tag = null, $isVersionKey = false) {
-		$key = ConfigManager::getConfig("Db", "Memcache")->AuxConfig->keyPrefix . ":";
+		$key = $this->config->keyPrefix . ":";
 
 		if ($isVersionKey) {
 			$key .= self::KEY_VERSION_PREFIX . ":";
